add TestDriver and Report

+++ lab08/prj/ModulesMelnik.h
@@ -1,6 +1,6 @@
 #ifndef MODULESMELNIK_H_INCLUDED
 #define MODULESMELNIK_H_INCLUDED
 
-float s_calculation(float x, float z);
+float s_calculation(int x, int z);
 
 #endif // MODULESMELNIK_H_INCLUDED
